🔖 Release 0.8.0

+++ src/main/java/com/github/antoinejt/jasc/calculator/OperationType.java
@@ -30,24 +30,24 @@ package com.github.antoinejt.jasc.calculator;
 import java.util.function.BiFunction;
 
 public enum OperationType {
-    ADDITION(Float::sum),
+    ADDITION(Double::sum),
     SUBSTRACTION((a, b) -> b - a),
     MULTIPLICATION((a, b) -> b * a),
     DIVISION((a, b) -> b / a),
     MODULO((a, b) -> b % a),
     POWER(OperationType::pow);
 
-    private final BiFunction<Float, Float, Float> operation;
+    private final BiFunction<Double, Double, Double> operation;
 
-    OperationType(BiFunction<Float, Float, Float> operation) {
+    OperationType(BiFunction<Double, Double, Double> operation) {
         this.operation = operation;
     }
 
-    private static float pow(float a, float b) {
-        return (float) Math.pow(b, a);
+    private static double pow(double a, double b) {
+        return Math.pow(b, a);
     }
 
-    public float apply(float a, float b) {
+    public double apply(double a, double b) {
         return operation.apply(a, b);
     }
 }
